arrangeDirectoryAndFixRoute
