+++ src/App.js
@@ -6,11 +6,11 @@ import {
   Redirect
 } from "react-router-dom";
 
-import Login from "./components/Login";
-import Logout from "./Logout";
-import Top from "./components/Top";
-import Favo from "./components/Favo";
-import Nearby from "./components/Nearby";
+import Login from "./components/pages/Login";
+import Logout from "./components/pages/Logout";
+import Target from "./components/Target";
+import Sns from "./components/Sns";
+import Notification from "./components/Notification";
 import Auth from "./Auth";
 
 import Header from "./components/organisms/Header";
@@ -28,9 +28,9 @@ export default function App() {
         <Auth>
           <Header />
           <Switch>
-            <Route exact path="/target" component={Top} />
-            <Route exact path="/sns" component={Favo} />
-            <Route exact path="/notification" component={Nearby} />
+            <Route exact path="/target" component={Target} />
+            <Route exact path="/sns" component={Sns} />
+            <Route exact path="/notification" component={Notification} />
             <Redirect from="/" to="/target" />
           </Switch>
           <Footer />
