+++ src/components/organisms/Footer.js
@@ -1,13 +1,13 @@
-import React, { useState } from "react";
+import React, { useEffect } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import { useHistory, useLocation } from "react-router-dom";
 
 //ButtomNavi
 import BottomNavigation from "@material-ui/core/BottomNavigation";
 import BottomNavigationAction from "@material-ui/core/BottomNavigationAction";
-import RestoreIcon from "@material-ui/icons/Restore";
-import FavoriteIcon from "@material-ui/icons/Favorite";
-import LocationOnIcon from "@material-ui/icons/LocationOn";
+import TargetIcon from "@material-ui/icons/CheckCircleOutline";
+import SnsIcon from "@material-ui/icons/Forum";
+import NotificationIcon from "@material-ui/icons/Notifications";
 
 const useStyles = makeStyles(theme => ({
   footer: {
@@ -16,8 +16,7 @@ const useStyles = makeStyles(theme => ({
     width: "100%",
     zIndex: "100",
     margin: "0",
-    padding: "0",
-    width: "100%"
+    padding: "0"
   }
 }));
 
@@ -25,7 +24,13 @@ function Footer() {
   const classes = useStyles();
   const history = useHistory();
   const locate = useLocation();
-  const [value, setValue] = React.useState(locate.pathname.slice(1));
+  const tab = locate.search.slice(1);
+  const [value, setValue] = React.useState(
+    tab ? tab : locate.pathname.slice(1)
+  );
+  useEffect(() => {
+    setValue(tab ? tab : locate.pathname.slice(1));
+  }, [locate, tab, locate.pathname]);
 
   return (
     <footer className={classes.footer}>
@@ -38,18 +43,14 @@ function Footer() {
         showLabels
       >
         <BottomNavigationAction
-          label="Recents"
-          icon={<RestoreIcon />}
+          label="Target"
+          icon={<TargetIcon />}
           value="target"
         />
+        <BottomNavigationAction label="SNS" icon={<SnsIcon />} value="sns" />
         <BottomNavigationAction
-          label="Favorites"
-          icon={<FavoriteIcon />}
-          value="sns"
-        />
-        <BottomNavigationAction
-          label="Nearby"
-          icon={<LocationOnIcon />}
+          label="Notification"
+          icon={<NotificationIcon />}
           value="notification"
         />
       </BottomNavigation>
