+++ src/components/pages/Login.js
@@ -11,9 +11,9 @@ import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
 import Container from "@material-ui/core/Container";
 import { useHistory } from "react-router-dom";
-import firebase from "../firebase";
-import googleSignin from "../img/googleSignin.png";
-import OverlaySpinner from "./OverlaySpinner";
+import firebase from "../utils/firebase";
+import googleSignin from "../../img/googleSignin.png";
+import OverlaySpinner from "./../utils/OverlaySpinner";
 
 const useStyles = makeStyles(theme => ({
   "@global": {
