+++ src/components/pages/Logout.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { Link } from "react-router-dom";
-import firebase from "./firebase";
+import firebase from "../utils/firebase";
 
 export default function Logout() {
   useEffect(() => {
